(PC-16310)[PRO] refactor: signup to implement scroll to error
