+++ pro/src/pages/Signup/SignupContainer/MaybeAppUserDialog/MaybeAppUserDialog.scss
