--- pro/src/pages/Signup/SignupForm/SignupForm.tsx
@@ -1,259 +1,0 @@
-import { Form, FormikProvider, useFormik } from 'formik'
-import React, { useEffect, useRef, useState } from 'react'
-import { useHistory } from 'react-router-dom'
-
-import { api } from 'apiClient/api'
-import { BannerInvisibleSiren, BannerRGS } from 'components/Banner'
-import FormLayout from 'components/FormLayout'
-import LegalInfos from 'components/LegalInfos/LegalInfos'
-import { Events } from 'core/FirebaseEvents/constants'
-import { getSirenDataAdapter } from 'core/Offerers/adapters'
-import useAnalytics from 'hooks/useAnalytics'
-import useLogEventOnUnload from 'hooks/useLogEventOnUnload'
-import { useModal } from 'hooks/useModal'
-import useNotification from 'hooks/useNotification'
-import useRedirectLoggedUser from 'hooks/useRedirectLoggedUser'
-import { Button, SubmitButton, TextInput, Checkbox } from 'ui-kit'
-import { ButtonVariant } from 'ui-kit/Button/types'
-import { PasswordInput, SirenInput } from 'ui-kit/form'
-import PhoneNumberInput from 'ui-kit/form/PhoneNumberInput'
-import { removeWhitespaces } from 'utils/string'
-
-import {
-  MAYBE_APP_USER_APE_CODE,
-  SIGNUP_FORM_DEFAULT_VALUES,
-} from './constants'
-import MaybeAppUserDialog from './MaybeAppUserDialog'
-import OperatingProcedures from './OperationProcedures'
-import { ISignupApiErrorResponse, ISignupFormValues } from './types'
-import { validationSchema } from './validationSchema'
-
-const SignupForm = (): JSX.Element => {
-  const history = useHistory()
-  const notification = useNotification()
-  const [showAnonymousBanner, setShowAnonymousBanner] = useState(false)
-  const { visible, showModal, hideModal } = useModal()
-  const { logEvent } = useAnalytics()
-  useRedirectLoggedUser()
-
-  useEffect(() => {
-    const script = document.createElement('script')
-
-    script.src = '//js.hs-scripts.com/5119289.js'
-    script.async = true
-    script.type = 'text/javascript'
-    script.id = 'hs-script-loader'
-    script.defer = true
-
-    document.body.appendChild(script)
-    return () => {
-      const script = document.getElementById('hs-script-loader') as Node
-      document.body.removeChild(script)
-    }
-  }, [])
-
-  const onSubmit = (values: ISignupFormValues) => {
-    const { legalUnitValues, ...flattenvalues } = values
-    const { firstName, siren } = flattenvalues
-    api
-      .signupPro({
-        ...flattenvalues,
-        ...legalUnitValues,
-        siren: removeWhitespaces(siren),
-        publicName: firstName,
-      })
-      .then(() => onHandleSuccess())
-      .catch(response => onHandleFail(response.errors ? response.errors : {}))
-  }
-
-  const onHandleSuccess = () => {
-    logEvent?.(Events.SIGNUP_FORM_SUCCESS, {})
-    history.replace('/inscription/confirmation')
-  }
-
-  const onHandleFail = (errors: ISignupApiErrorResponse) => {
-    for (const field in errors) {
-      formik.setFieldError(field, (errors as any)[field])
-    }
-
-    notification.error(
-      'Une ou plusieurs erreurs sont présentes dans le formulaire.'
-    )
-    formik.setSubmitting(false)
-  }
-
-  const formik = useFormik({
-    initialValues: SIGNUP_FORM_DEFAULT_VALUES,
-    onSubmit: onSubmit,
-    validationSchema,
-    validateOnChange: false,
-  })
-
-  const getSirenAPIData = async (siren: string) => {
-    setShowAnonymousBanner(false)
-    const response = await getSirenDataAdapter(siren)
-    if (response.isOk) {
-      let values = undefined
-      if (response.payload.values != null) {
-        const { apeCode, ...rest } = response.payload.values
-        values = rest
-        if (MAYBE_APP_USER_APE_CODE.includes(apeCode)) {
-          showModal()
-        }
-      }
-
-      formik.setFieldValue('legalUnitValues', values)
-    } else {
-      formik.setFieldError('siren', response.message)
-      if (
-        response.message ==
-        'Les informations relatives à ce SIREN ou SIRET ne sont pas accessibles.'
-      ) {
-        setShowAnonymousBanner(true)
-      }
-    }
-  }
-
-  // Track the state of the form when the user gives up
-  const touchedRef = useRef(formik.touched)
-  const errorsRef = useRef(formik.errors)
-
-  useEffect(() => {
-    touchedRef.current = formik.touched
-    errorsRef.current = formik.errors
-  }, [formik.touched, formik.errors])
-
-  const logFormAbort = (): void | undefined => {
-    const filledFields = Object.keys(touchedRef.current)
-    if (filledFields.length === 0) {
-      return
-    }
-    // formik.errors contains every fields with errors even if they have not been touched.
-    // We filter theses errors by touched fields to only have fields filled by the user with errors
-    const filledWithErrors = Object.keys(
-      Object.fromEntries(
-        Object.entries(errorsRef.current).filter(([errorKey]) =>
-          filledFields.includes(errorKey)
-        )
-      )
-    )
-    return logEvent?.(Events.SIGNUP_FORM_ABORT, {
-      filled: filledFields,
-      filledWithErrors: filledWithErrors,
-    })
-  }
-
-  // Track the form state on tab closing
-  useLogEventOnUnload(() => logFormAbort())
-
-  // Track the form state on component unmount
-  useEffect(() => {
-    return () => {
-      if (Object.entries(errorsRef.current).length === 0) {
-        return
-      }
-      logFormAbort()
-    }
-  }, [])
-  return (
-    <section className="sign-up-form-page">
-      {visible && <MaybeAppUserDialog onCancel={hideModal} />}
-      <div className="content">
-        <h1>Créer votre compte professionnel</h1>
-        <OperatingProcedures />
-
-        <div className="sign-up-tips">
-          Tous les champs sont obligatoires sauf mention contraire
-        </div>
-        <FormikProvider value={formik}>
-          <Form onSubmit={formik.handleSubmit}>
-            <FormLayout>
-              <div className="sign-up-form">
-                <FormLayout.Row>
-                  <TextInput
-                    label="Adresse e-mail"
-                    name="email"
-                    placeholder="mail@exemple.com"
-                  />
-                </FormLayout.Row>
-                <FormLayout.Row>
-                  <PasswordInput
-                    name="password"
-                    label="Mot de passe"
-                    placeholder="Votre mot de passe"
-                    withErrorPreview={true}
-                  />
-                </FormLayout.Row>
-                <FormLayout.Row>
-                  <TextInput
-                    label="Nom"
-                    name="lastName"
-                    placeholder="Votre nom"
-                  />
-                </FormLayout.Row>
-                <FormLayout.Row>
-                  <TextInput
-                    label="Prénom"
-                    name="firstName"
-                    placeholder="Votre prénom"
-                  />
-                </FormLayout.Row>
-                <FormLayout.Row>
-                  <PhoneNumberInput
-                    name="phoneNumber"
-                    label={
-                      'Téléphone (utilisé uniquement par l’équipe du pass Culture)'
-                    }
-                  />
-                </FormLayout.Row>
-                <div className="siren-field">
-                  <FormLayout.Row>
-                    <SirenInput
-                      label="SIREN de la structure que vous représentez"
-                      onValidSiren={getSirenAPIData}
-                    />
-                  </FormLayout.Row>
-                  <span className="field-siren-value">
-                    {formik.values.legalUnitValues.name}
-                  </span>
-                  {showAnonymousBanner && <BannerInvisibleSiren />}
-                </div>
-                <FormLayout.Row>
-                  <Checkbox
-                    hideFooter
-                    label="J’accepte d’être contacté par e-mail pour recevoir les
-                      nouveautés du pass Culture et contribuer à son
-                      amélioration (facultatif)"
-                    name="contactOk"
-                  />
-                </FormLayout.Row>
-                <LegalInfos
-                  className="sign-up-infos-before-signup"
-                  title="Créer mon compte"
-                />
-              </div>
-              <div className="buttons-field">
-                <Button
-                  onClick={() => history.push('/connexion')}
-                  variant={ButtonVariant.SECONDARY}
-                >
-                  J’ai déjà un compte
-                </Button>
-                <SubmitButton
-                  className="primary-button"
-                  isLoading={formik.isSubmitting}
-                  disabled={!formik.dirty || !formik.isValid}
-                >
-                  Créer mon compte
-                </SubmitButton>
-              </div>
-              <BannerRGS />
-            </FormLayout>
-          </Form>
-        </FormikProvider>
-      </div>
-    </section>
-  )
-}
-
-export default SignupForm
